Do not run 'set' disposes for undefined
The "overeager" type guarding that caused the bug fixed in the previous
commit, was it turns out *also* preventing spurious `dispose()` method
calls the first time an entry was added to the cache.

That is now fixed, by checking not only that `noDisposeOnSet` is not
set, but also that the item is actually in the cache prior to being
added.